requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
distro==1.9.0
    # via openai
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
jiter==0.9.0
    # via openai
openai==1.76.0
    # via -r requirements.in
pydantic==2.11.3
    # via
    #   -r requirements.in
    #   openai
pydantic-core==2.33.1
    # via pydantic
pyperclip==1.9.0
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
typing-extensions==4.13.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
